Renamed utills.js to widgets.js and added nicer loading widget
