+++ frontend/src/features/authenticated-app/AuthenticatedApp.js
@@ -2,12 +2,14 @@ import React from "react";
 import { useCourseSetup } from "../../common/api";
 import SetupScreen from "../setup-course/SetupScreen";
 import UploadScreen from "../upload-exams/UploadScreen";
+import { LoadingPage } from "../widgets";
 
 export default function AuthenticatedApp({ courseId }) {
   const query = useCourseSetup(courseId);
 
   if (query.isLoading) {
     return <div>Loading...</div>;
+    return <LoadingPage>Loading...</LoadingPage>;
   }
 
   if (query.isError) {
